Seventh Update

+++ AptManager_FinalProject/MongoEntites/CreditCard.cs
@@ -17,6 +17,6 @@ namespace Models.MongoEntites
         public string CardNumber { get; set; }
         public string CVC { get; set; }
         public string ExpireDate { get; set; }
-        public decimal Balance { get; set; } = 2000;
+        public decimal Balance { get; set; } = (decimal)2000;
     }
 }
